Split cut sound patches up...
and use injected NoteCutSoundEffectManager to change the cut sounds
